Add support for distinguishing forceStop() and reset()
This CL renames stop() to forceStop() and marks it as private.
forceStop() is now only called by the SupportTimeoutAlarm. After a call
to forceStop() any call to start() is ignored without a prior call to
reset().

IpClient calls reset() in RunningState#exit() to "prime" the address
registration logic to be restarted.

This solves a potential issue where the SupportTimeoutAlarm fires, but a
subsequent call to start() (because an RTM_NEWLINK with M or O bit set
was received) would start the tracker again, even though the network
does not support address registration. Instead of keeping state in
IpClient or IpClientLinkObserver, this change encapsulates all the state
inside Dhcp6AddrRegTracker. IpClient can continue calling start()
multiple times.

On a side note: I have also noticed a potential issue with
SupportTimeoutAlarm. It seems to suffer from the classic handler
callback race. In particular, it is currently possible (though somewhat
unlikely) for SupportTimeoutAlarm to fire right as IpClient restarts. In
that case, it might(?) be possible for the alarm execution to be delayed
until the IpClient has restarted.
Though, in reality, the AddrRegTracker gets started in the context of
an LinkProperties changed callback, which -- I believe -- must be
ordered behind the stale SupportTimeoutAlarm event. So as it stands,
this might not actually be an issue. In any case it is something to
watch out for and might be worth fixing. In a perfect world, the
AddrRegTracker would not be making any assumptions on IpClient handler
interactions.

Test: Dhcp6AddrRegTrackerTest
Flag: EXEMPT mainline
Change-Id: I86a5e3ebfb87bda0756aaf503380d49cba4dfcfc