Enable flag-based tests in IpClientRootTest
There are two types of test annotations which can be used with flags:
- FlagsEnabled / FlagsDisabled: These annotations write the flag value
  in process before running the test.
- RequiresFlagsEnabled / RequiresFlagsDisabled: These annotations skip
  the test based on the actual flag value.

However, the existing test rules prevent the use of both annotations at
on the same test.

The IpClient integration test setup is special in the sense that the
same set of tests are run in-process (IpClientSignatureTest) and
out-of-process (IpClientRootTest). In this case, the desired behavior is
for the in-process tests to set the flag value and for the
out-of-process tests to read the flag value and optionally skip the
test.

This CL adds a new test rule which closely matches the existing
CheckFlagsRule, but instead of considering RequiresFlagsEnabled /
RequiresFlagsDisabled annotations, it looks at EnableFlags /
DisableFlags instead. This way, the existing SetFlagsRule (now moved to
IpClientSignatureTest) can be reused.

Test: TH
Flag: NONE yes
Change-Id: I83c25e66b0a6e80a4e828669cfb927eaf76dfaf6