Add terrible error metrics in ClearingClearingIpAddressesState
Since mLinkProperties is reset in StoppedState#enter, we think
readyToProceed will always return true and transition directly to
RunningState.

This check was originally intended to catch abnormal exits (e.g.,
tethering exits) where an IP address might remain on the interface.
Given that tethering should now operate on a separate interface with
IpClient, this issue should be obsolete.

Add a terrible error metric for this rare race condition, to detect
and report if this can actually occur.

Flag: NONE yes
Bug: 240616892
Test: TH
Change-Id: I165e76a2de89e572b59a41fa34ddd93f5fa678e5