Fix the testKernelDeletesIPv6AddressesOnValidLifetimeExpires.
The test case was intended to trigger `onProvisioningFailure` after
the link address's valid lifetime expires, and all global IPv6
addresses will be removed from the interface.

However, when the IPv6 address registration beta flag is enabled, a
few things change:

- actual lifetime populated: The link address's lifetime is populated
  with the actual lifetime.
- address depreciation: The link address eventually becomes deprecated,
  causing `LinkAddress.isPreferred` to return false.
- callback blocked: This results in the provisioning delta being
  `PROV_CHANGE_STILL_NOT_PROVISIONED`, which prevents
  `onProvisioningFailure` from being triggered.

Instead verify if the `onProvisioningFailure` will be called, verify if
all global IPv6 addresses will be deleted eventually, this is regardless
of the beta flags.

Additionally, we observed that `LinkProperties.equals()` currently
ignores `expirationTime` and `deprecationTime`. This means that address
flag updates (like moving from preferred to deprecated) might be
considered "the same" for `LinkProperties`, potentially blocking
callbacks. This specific behavior requires further investigation,
though it is not the primary focus of this fix.

Bug: 445845002
Test: TH
Flag: NONE yes
Change-Id: I27754f7d502950a7ed151fc41d523d4e0a499a98