Report capport data on successful API probe
Always report capport data and return it  when the Capport API probe
succeeds. This addresses an edge case where an initial HTTP probe with
capport data succeeded, but a subsequent fallback probe failed and
detected a portal. In such a scenario, the capport data should still be
reported and used.

Test: TH
Flag: EXEMPT BUGFIX
Change-Id: I76d207465193a29126cdc32f23de959654abda63