Fix IpClientLinkObserver to receive RTM_NEWLINK messages triggered by RA
The old code incorrectly ignored all RTM_NEWLINK messages with
ifinfomsg.family set to AF_INET6. However, RTM_NEWLINK messages
triggered by an incoming RA do set family to AF_INET6.

This code optionally checks for AF_INET6 (depending on addr reg flag
value) and processes the flags. In order to prevent putting a bunch of
unnecessary messages on the handler, it only checks for the M or O bit
if the flags have changed.

Note that the flag is checked inside IpClient and that
startDhcp6AddrReg() posts a message which is effectively a noop to the
handler if the flag is not enabled. We could consider moving that flag
check inside of processInet6Flags instead, so the entire code path is
flagged off.

Test: TH
Flag: NONE yes
Change-Id: I1b5c287485e5ddf3a8ae258629077702a18132c5