Add APF counters for low power standby port exemption
Add counters:
- DROPPED_LOW_POWER_STANDBY
- PASSED_LOW_POWER_STANDBY_PORT_ALLOWED

When in low power standby, all packets are dropped, except for
particular UDP or TCP packets with an allowlisted destination port.
Therefore, a DROPPED_LOW_POWER_STANDBY packet might be have a
destination port that is not allowlisted, or it might not be a UDP or
TCP packet.

Bug: 408247593
Test: atest NetworkStackTests
Flag: NONE yes
Change-Id: I57184d14e21ae562527ccce883210ab89c0cb430