Wait for NetworkStack connector to be alive
In cases where the NetworkStack and system_server crashed and restarted,
the "old" NetworkStack connector will not be removed from the
ServiceManager and may be returned by NetworkStack.getService().

Ensure that the connector is alive before stopping polling in
ModuleNetworkStackClient.

This is similar to change ID: I549d5a07966ebaed7e4a4ebe96e831dcac3b05bd
but for NetworkStack instead of the tethering connector.

This patch also refactors the test to remove the dependency on
mockito-extended, as it uses heavy instrumentation that sometimes causes
issues in tests.

Bug: 461838689
Test: atest
Flag: EXEMPT BUGFIX
Change-Id: Iee8b4b7d3046ef9c0facee9113c0898f6ed8dfdd